requirements.txt
annotated-types==0.6.0
anyio==4.0.0
beautifulsoup4==4.12.2
certifi==2023.7.22
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
sniffio==1.3.0
soupsieve==2.5
typing_extensions==4.8.0
